Move curve_trees.h type defs into fcmp_pp_types.h
