+++ src/blockchain_db/blockchain_db.h
@@ -588,7 +588,7 @@ private:
 
   virtual std::vector<crypto::ec_point> grow_with_tree_extension(const fcmp_pp::curve_trees::CurveTreesV1::TreeExtension &tree_extension) = 0;
 
-  virtual fcmp_pp::curve_trees::PathBytes get_path(const fcmp_pp::curve_trees::PathIndexes &path_indexes) const = 0;
+  virtual fcmp_pp::PathBytes get_path(const fcmp_pp::PathIndexes &path_indexes) const = 0;
 
   virtual uint64_t find_leaf_idx_by_unified_id_bounded_search(uint64_t unified_id, uint64_t leaf_idx_start, uint64_t leaf_idx_end) const = 0;
 
@@ -1869,12 +1869,12 @@ public:
 
   void trim_tree(const uint64_t new_n_leaf_tuples, const uint64_t trim_block_idx);
 
-  std::pair<uint64_t, fcmp_pp::curve_trees::PathBytes> get_last_path(const uint64_t block_idx) const;
+  std::pair<uint64_t, fcmp_pp::PathBytes> get_last_path(const uint64_t block_idx) const;
 
   uint64_t get_path_by_unified_id(const std::vector<uint64_t> &unified_ids,
     const uint64_t as_of_n_blocks,
     std::vector<uint64_t> &leaf_idxs_out,
-    std::vector<fcmp_pp::curve_trees::PathBytes> &paths_out) const;
+    std::vector<fcmp_pp::PathBytes> &paths_out) const;
 
   /**
    * @brief add outs to locked outputs tables
